+++ NEWS.md
@@ -1,6 +1,7 @@
 # DataEditR 0.1.4
 
 * Add support for writing data to `googlesheets` by calling `write_fun` as the application closes.
+* Fix `data_edit()` to ensure that we don't search for data through RStudio addin if RStudio is not in use. 
 
 # DataEditR 0.1.3